fixed types and imports
