+++ src/common/components/typography/index.ts
@@ -1,0 +1,2 @@
+export { Typography } from './Typography'
+export { typographyVariants} from './typographyVariants'
